wifi_presence_detection/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.2.0
chardet==3.0.4
connect-box==0.2.5
defusedxml==0.6.0
idna==2.8
multidict==4.5.2
python-nmap==0.6.1
yarl==1.3.0
